% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{autoplot.imputed_trait}
\alias{autoplot.imputed_trait}
\title{Summarise Bootstrap traits}
\usage{
\method{autoplot}{imputed_trait}(imputed_traits)
}
\arguments{
\item{BootstrapMoments}{trait moments from trait_np_bootstrap}
}
\value{

}
\description{
Bootstraptraits


}
