% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{autoplot.selected_trait}
\alias{autoplot.selected_trait}
\title{Coverage plot of selected Traits}
\usage{
\method{autoplot}{selected_trait}(selected_traits, other_col_how, ...)
}
\arguments{
\item{selected_traits}{output from trait_np_bootstrap or trait_parametric_bootstrap
funtion.}

\item{other_col_how}{what to do with the other columns in other data.
 Options are to `filter` by one of the columns, add them to the x-`axis`,
`facet` by them, or to `ignore`.}

\item{...}{optional filters for use with `other_col_how = "filter"`}
}
\value{
a ggplot
}
\description{
Function calculates the trait coverage of the community 
for each level of the sampling hierarchy and makes a barplot.

Shows at which level the data are coming from in each plot.
}
\examples{
require("ggplot2")
data(community)
data(trait)
selected_traits <-trait_select(comm = community, traits = trait,
                 scale_hierarchy = c("Site", "PlotID"),
                 taxon_col = "Taxon", value_col = "Value",
                 trait_col = "Trait", abundance_col = "Cover")
autoplot(selected_traits)
}
