% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_to_df.R
\name{cor_to_df}
\alias{cor_to_df}
\title{Correlation to dataframe}
\usage{
cor_to_df(corr)
}
\arguments{
\item{corr}{correlation matrix}
}
\value{
A data.frame of correlations
}
\description{
Helper function for bootstrap_traits_multivariate that
extracts results from a correlation matrix
}
\examples{
x <- matrix(ncol = 5, rnorm(20))
colnames(x) <- letters[1:5]
cor(x) |> cor_to_df()

}
\references{
Modified from https://stackoverflow.com/a/23476844/2055765
}
