% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_handler.R
\name{distribution_handler}
\alias{distribution_handler}
\title{Get different distributions}
\usage{
distribution_handler(parm1, parm2, n = 1, type)
}
\arguments{
\item{parm1}{the first parameter in a fitted distribution}

\item{parm2}{the second parameter in a fitted distribution}

\item{n}{Number of draws.  Defaults to 1}

\item{type}{Distribution type to fit.
One of "normal", "lognormal", or "beta"}
}
\value{
numeric vector of length \code{n} from required distribution.
}
\description{
Internal helper function for drawing traits
}
\keyword{internal}
