% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trait_distributions.R
\name{fit_trait_distributions}
\alias{fit_trait_distributions}
\title{Fit trait distributions}
\usage{
fit_trait_distributions(imputed_traits, distribution_type = "normal")
}
\arguments{
\item{imputed_traits}{imputed trait and community data
in long format}

\item{distribution_type}{the type of statistical distribution to use.
Character or named list. Currently accepts "normal","lognormal", and "beta".}
}
\value{
a tibble
}
\description{
Fits distributions to imputed traits and
returns a tibble containing the fitted parameters.


}
\note{
Distribution type should be a single character (e.g. "normal")
or a named list (e.g. list(height="normal",mass="lognormal"))
}
\examples{
data(community)
data(trait)
imputed_traits <-trait_impute(comm = community, traits = trait,
                 scale_hierarchy = c("Site", "PlotID"),
                 taxon_col = "Taxon", value_col = "Value",
                 trait_col = "Trait", abundance_col = "Cover")
fitted_distributions <- fit_trait_distributions(imputed_traits = imputed_traits,
distribution_type = "normal")

}
