% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_filled_trait.R
\name{fortify.filled_trait}
\alias{fortify.filled_trait}
\title{Fortify filled Traits}
\usage{
\method{fortify}{filled_trait}(filled_traits, other_col_how, ...)
}
\arguments{
\item{filled_traits}{filled traits from trait_fill}

\item{other_col_how}{what to do with the other columns in other data.
Options are to \code{filter} by one of the columns, add them to the x-\code{axis},
\code{facet} by them, or to \code{ignore}.}

\item{...}{optional filters for use with \code{other_col_how = "filter"}}
}
\value{
a tibble
}
\description{
Calculates at which level the data are coming from in each plot.
}
