% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_impute_traits.R
\name{fortify.imputed_trait}
\alias{fortify.imputed_trait}
\title{Fortify Imputed Traits}
\usage{
\method{fortify}{imputed_trait}(imputed_traits, other_col_how, ...)
}
\arguments{
\item{imputed_traits}{imputed traits from trait_impute}

\item{other_col_how}{what to do with the other columns in other data.
Options are to `filter` by one of the columns, add them to the x-`axis`,
`facet` by them, or to `ignore`.}

\item{...}{optional filters for use with `other_col_how = "filter"`}
}
\value{
a tibble
}
\description{
Calculates at which level the data are coming from in each plot.
}
