% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_select_traits.R
\name{fortify.selected_trait}
\alias{fortify.selected_trait}
\title{Fortify selected Traits}
\usage{
\method{fortify}{selected_trait}(selected_traits, other_col_how, ...)
}
\arguments{
\item{selected_traits}{selected traits from trait_select}

\item{other_col_how}{what to do with the other columns in other data.
Options are to `filter` by one of the columns, add them to the x-`axis`,
`facet` by them, or to `ignore`.}

\item{...}{optional filters for use with `other_col_how = "filter"`}
}
\value{
a tibble
}
\description{
Calculates at which level the data are coming from in each plot.
}
