% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_filled_trait.R
\name{fortify_filled_trait}
\alias{fortify_filled_trait}
\title{Fortify filled Traits}
\usage{
fortify_filled_trait(object, other_col_how, ...)
}
\arguments{
\item{object}{filled traits from \code{trait_fill()}.}

\item{other_col_how}{what to do with the other columns in other data.
Options are to \code{filter} by one of the columns, add them to the x-\code{axis},
\code{facet} by them, or to \code{ignore}.}

\item{...}{optional filters for use with \code{other_col_how = "filter"}}
}
\value{
a tibble
}
\description{
Calculates at which level the data are coming from in each plot.
}
