% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{Get confidence intervals}
\usage{
get_ci(data, sd_mult = 1, ci = 0.95, which, parametric = T)
}
\arguments{
\item{data}{Numeric vector}

\item{sd_mult}{SD multiplier.  defaults to 1}

\item{ci}{0 to 1}

\item{which}{"high" or "low"}

\item{parametric}{if T, sd used, if F, quantile}
}
\description{
Internal helper function for generating CIs.
}
\keyword{internal}
