% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Do something quietly}
\usage{
quiet(x)
}
\arguments{
\item{x}{Code to be executed quietly}
}
\description{
Internal helper function
used to prevent printed messages
}
\keyword{internal}
