% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{trait_bootstrap}
\alias{trait_bootstrap}
\title{Bootstrap traits}
\usage{
trait_bootstrap(traits_comm, nrep = 100, sample_size = 200)
}
\arguments{
\item{traits_comm}{imputed trait and community data in long format}

\item{nrep}{number of bootstrap replicates}

\item{sample_size}{number of plants per sample}
}
\value{

}
\description{
Bootstrap impute traits


}
