% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_traits.R
\name{trait_np_bootstrap}
\alias{trait_np_bootstrap}
\title{Bootstrap traits}
\usage{
trait_np_bootstrap(imputed_traits, nrep = 100, sample_size = 200, raw = FALSE)
}
\arguments{
\item{imputed_traits}{imputed trait and community data in long format}

\item{nrep}{number of bootstrap replicates}

\item{sample_size}{number of plants per sample}

\item{raw}{logical argument to get the raw data of the distributions}
}
\value{
a tibble
}
\description{
Bootstrap impute traits


}
\examples{
data(community)
data(trait)
imputed_traits <-trait_impute(comm = community, traits = trait,
                 scale_hierarchy = c("Site", "PlotID"),
                 taxon_col = "Taxon", value_col = "Value",
                 trait_col = "Trait", abundance_col = "Cover")
boot_traits <- trait_np_bootstrap(imputed_traits)
}
