% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_bootstrap_moments.R
\name{trait_summarise_boot_moments}
\alias{trait_summarise_boot_moments}
\title{Summarise Bootstrap traits}
\usage{
trait_summarise_boot_moments(bootstrap_moments)
}
\arguments{
\item{bootstrap_moments}{trait moments from trait_np_bootstrap}
}
\value{
tibble with the grouping variables and the mean of each moment (+/- 1SD)
}
\description{
Bootstraptraits


}
