% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{autoplot.imputed_trait}
\alias{autoplot.imputed_trait}
\title{Coverage plot of Imputed Traits}
\usage{
\method{autoplot}{imputed_trait}(imputed_traits)
}
\arguments{
\item{imputed_traits}{trait moments from trait_np_bootstrap}
}
\value{
a ggplot
}
\description{
Shows at which level the data are coming from in each plot.
}
