% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community}
\alias{community}
\title{Community data}
\format{
A data frame with 110 rows and 4 variables:
\describe{
\item{Taxon}{species name}
\item{Cover}{cover, in percent}
\item{Site}{site name}
\item{PlotID}{plot name}
}
}
\source{
\url{https://www.uib.no/en/rg/EECRG/114808/plant-functional-traits-course-4}
}
\usage{
community
}
\description{
A dataset containing plant cover in control plots on Svalbard from PFCT4
TraitTrain course.
}
\keyword{datasets}
