% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{fortify.imputed_trait}
\alias{fortify.imputed_trait}
\title{Fortify Imputed Traits}
\usage{
\method{fortify}{imputed_trait}(imputed_traits)
}
\arguments{
\item{imputed_traits}{trait moments from trait_np_bootstrap}
}
\value{
a tibble
}
\description{
Calculates at which level the data are coming from in each plot.
}
