% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{SummariseBootMoments}
\alias{SummariseBootMoments}
\title{Summarise Bootstrap traits}
\usage{
SummariseBootMoments(BootstrapMoments)
}
\arguments{
\item{BootstrapMoments}{trait moments from trait_np_bootstrap}
}
\value{

}
\description{
Bootstraptraits


}
