% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait}
\alias{trait}
\title{Trait data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Taxon}{species name}
  \item{Site}{site name}
  \item{PlotID}{plot name}
  \item{Trait}{trait name with unit}
  \item{Value}{trait value}
}
}
\source{
\url{http://https://www.uib.no/en/rg/EECRG/114808/plant-functional-traits-course-4}
}
\usage{
trait
}
\description{
A dataset containing plant traits in control plots on valbard from PFCT4 TraitTrain course.
}
\keyword{datasets}
