% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait}
\alias{trait}
\title{Trait data}
\format{
A data frame with 705 rows and 6 variables:
\describe{
\item{Site}{site name}
\item{PlotID}{plot name}
\item{Taxon}{species name}
\item{ID}{Unique leaf ID}
\item{Trait}{trait name with unit}
\item{Value}{trait value}
}
}
\source{
\url{https://www.uib.no/en/rg/EECRG/114808/plant-functional-traits-course-4}
}
\usage{
trait
}
\description{
A dataset containing plant traits in control plots on Svalbard from PFCT4
TraitTrain course.
}
\keyword{datasets}
