% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_missing.R
\name{trait_missing}
\alias{trait_missing}
\title{Which taxa lack traits}
\usage{
trait_missing(selected_trait, comm)
}
\arguments{
\item{selected_trait}{output of trait_select function.}

\item{comm}{community data}
}
\description{
Function gives overview of which taxa are missing traits.
}
