% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_impute.R
\name{trait_np_bootstrap}
\alias{trait_np_bootstrap}
\title{Bootstrap traits}
\usage{
trait_np_bootstrap(imputed_traits, nrep = 100, sample_size = 200)
}
\arguments{
\item{imputed_traits}{imputed trait and community data in long format}

\item{nrep}{number of bootstrap replicates}

\item{sample_size}{number of plants per sample}
}
\value{
a tibble
}
\description{
Bootstrap impute traits


}
