% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_traits.R
\name{trait_np_bootstrap}
\alias{trait_np_bootstrap}
\title{Bootstrap traits}
\usage{
trait_np_bootstrap(selected_traits, nrep = 100, sample_size = 200, raw = FALSE)
}
\arguments{
\item{selected_traits}{output from the trait_select function.}

\item{nrep}{number of bootstrap replicates}

\item{sample_size}{bootstrap size}

\item{raw}{logical; argument to extract the raw data of the trait distributions.
The default is raw = FALSE. If raw = TRUE, nrep is restricted to 1 to avoid
memory issues.}
}
\value{
a tibble
}
\description{
Function for nonparametric bootstrap resampling to calculate community
weighted trait mean and higher moments.
}
\details{
The observed traits are re-sampled in proportion to their weights,
e.g. the abundance of a species or the biomass. Values across all individuals
in a community are resampled n times (nrep) to incorporate the full
spectrum of trait variation, generating n number (sample_size) of trait distributions.
From these distributions the function estimates the mean and the higher moments
including variance, skewness and kurtosis.

#' The output of trait_np_bootstrap can be summarized using
trait_summarize_boot_moments.
}
\examples{
data(community)
data(trait)
selected_traits <- trait_select(
  comm = community, traits = trait,
  scale_hierarchy = c("Site", "PlotID"),
  taxon_col = "Taxon", value_col = "Value",
  trait_col = "Trait", abundance_col = "Cover"
)
boot_traits <- trait_np_bootstrap(selected_traits)
}
