% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_traits_parametric.R
\name{trait_parametric_bootstrap}
\alias{trait_parametric_bootstrap}
\title{Bootstrap traits parametrically}
\usage{
trait_parametric_bootstrap(fitted_distributions, nrep = 100, sample_size = 200)
}
\arguments{
\item{fitted_distributions}{Fitted distribution object returned by trait_fit_distributions}

\item{nrep}{number of bootstrap replicates}

\item{sample_size}{number of plants per sample}
}
\value{
a tibble
}
\description{
Bootstrap fitted distributions


}
\examples{
data(community)
data(trait)
imputed_traits <-trait_impute(comm = community, traits = trait,
                 scale_hierarchy = c("Site", "PlotID"),
                 taxon_col = "Taxon", value_col = "Value",
                 trait_col = "Trait", abundance_col = "Cover")
fitted_distributions <- trait_fit_distributions(imputed_traits =
imputed_traits, distribution_type = "normal")
parametric_dists <- trait_parametric_bootstrap(fitted_distributions =
fitted_distributions, nrep = 10, sample_size = 100)
moment_summary <-
trait_summarise_boot_moments(bootstrap_moments = parametric_dists,
parametric = FALSE)
}
